fix(autocluster): avoid stopping autocluster too eagerly
Specifically when discovery found no nodes. This can happen due to a lot
of various race conditions in real-world systems, e.g.: DNS information
propagating with some delay.

Now the assumption is that discovery should find at least a single node.
With singleton clusters, this will be the current node most of time.